Add prettier formatter and format solidity code (#544)
* Add prettier formatter and format solidity code

* Add format script

* Improve format and slither scripts and update README.md

* Regenerate bindings

* Fix test coverage error and use node container

* Add openssh-client to node container

* Fix test coverage for licence.sol

* Add git to format_check job

* Revert build_check changes

+++ contracts/internals/burner.sol
@@ -18,8 +18,9 @@
 
 pragma solidity ^0.5.15;
 
+
 // The BurnerToken interface is the interface to a token contract which
 // provides the total burnable supply for the TokenHolder contract.
 interface IBurner {
-    function currentSupply() external view returns (uint);
+    function currentSupply() external view returns (uint256);
 }
