+++ contracts/mocks/token.sol
@@ -2,22 +2,22 @@ pragma solidity ^0.5.15;
 
 import "../externals/SafeMath.sol";
 
+
 /// @title Token is a mock ERC20 token used for testing.
 contract Token {
-
     using SafeMath for uint256;
 
-    event Approval(address indexed owner, address indexed spender, uint value);
+    event Approval(address indexed owner, address indexed spender, uint256 value);
     event Transfer(address indexed from, address indexed to, uint256 amount);
     /// @dev Total supply of tokens in circulation.
-    uint public totalSupply;
+    uint256 public totalSupply;
 
     /// @dev Balances for each account.
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
 
     /// @dev Transfer a token. This throws on insufficient balance.
-    function transfer(address to, uint amount) public returns (bool) {
+    function transfer(address to, uint256 amount) public returns (bool) {
         require(balanceOf[msg.sender] >= amount);
         balanceOf[msg.sender] -= amount;
         balanceOf[to] += amount;
@@ -25,11 +25,11 @@ contract Token {
         return true;
     }
 
-    function transferFrom(address _from, address _to, uint _value) public returns (bool) {
+    function transferFrom(address _from, address _to, uint256 _value) public returns (bool) {
         if (_to == address(0)) return false;
         if (balanceOf[_from] < _value) return false;
 
-        uint allowed = allowance[_from][msg.sender];
+        uint256 allowed = allowance[_from][msg.sender];
         if (allowed < _value) return false; //PROBLEM!
         /* require(_value <= allowed, "amount exceeds allowance"); */
 
@@ -40,7 +40,7 @@ contract Token {
         return true;
     }
 
-    function approve(address _spender, uint _value) public returns (bool) {
+    function approve(address _spender, uint256 _value) public returns (bool) {
         //require user to set to zero before resetting to nonzero
         if ((_value != 0) && (allowance[msg.sender][_spender] != 0)) {
             return false;
@@ -52,14 +52,14 @@ contract Token {
     }
 
     /// @dev Credit an address.
-    function credit(address to, uint amount) public returns (bool) {
+    function credit(address to, uint256 amount) public returns (bool) {
         balanceOf[to] += amount;
         totalSupply += amount;
         return true;
     }
 
     /// @dev Debit an address.
-    function debit(address from, uint amount) public {
+    function debit(address from, uint256 amount) public {
         balanceOf[from] -= amount;
         totalSupply -= amount;
     }
